move rawtextparser to parser folder #660
