--- analysis/import/RawTextParser/src/main/kotlin/de/maibornwolff/codecharta/importer/rawtextparser/model/BoolIntConversion.kt
@@ -1,4 +1,0 @@
-package de.maibornwolff.codecharta.importer.rawtextparser.model
-
-fun Boolean.toInt() = if (this) 1 else 0
-fun Int.toBool() = this == 1
