--- analysis/import/RawTextParser/src/main/kotlin/de/maibornwolff/codecharta/importer/rawtextparser/model/FileMetrics.kt
@@ -1,10 +1,0 @@
-package de.maibornwolff.codecharta.importer.rawtextparser.model
-
-class FileMetrics {
-    val metricMap = mutableMapOf<String, Double>()
-
-    fun addMetric(name: String, value: Int): FileMetrics {
-        metricMap[name] = value.toDouble()
-        return this
-    }
-}
