+++ analysis/parser/RawTextParser/src/main/kotlin/de/maibornwolff/codecharta/parser/rawtextparser/model/BoolIntConversion.kt
@@ -1,0 +1,4 @@
+package de.maibornwolff.codecharta.parser.rawtextparser.model
+
+fun Boolean.toInt() = if (this) 1 else 0
+fun Int.toBool() = this == 1
