+++ analysis/parser/RawTextParser/src/main/kotlin/de/maibornwolff/codecharta/parser/rawtextparser/model/FileMetrics.kt
@@ -1,0 +1,10 @@
+package de.maibornwolff.codecharta.parser.rawtextparser.model
+
+class FileMetrics {
+    val metricMap = mutableMapOf<String, Double>()
+
+    fun addMetric(name: String, value: Int): FileMetrics {
+        metricMap[name] = value.toDouble()
+        return this
+    }
+}
