+++ analysis/parser/RawTextParser/src/test/kotlin/de/maibornwolff/codecharta/indentationlevelparser/model/FileMetricsTest.kt
@@ -1,0 +1,18 @@
+package de.maibornwolff.codecharta.rawtextparser.model
+
+import de.maibornwolff.codecharta.parser.rawtextparser.model.FileMetrics
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Test
+
+class FileMetricsTest {
+    @Test
+    fun `Adding metric sould add metric to map`() {
+        val fileMetric = FileMetrics()
+        val result = mutableMapOf<String, Double>("foo" to 1.0, "bar" to 2.0)
+
+        fileMetric.addMetric("foo", 1)
+        fileMetric.addMetric("bar", 2)
+
+        Assertions.assertThat(fileMetric.metricMap).isEqualTo(result)
+    }
+}
