+++ analysis/parser/RawTextParser/src/test/resources/cc_projects/project_2.cc.json
