+++ analysis/parser/RawTextParser/src/test/resources/cc_projects/project_3.cc.json
