+++ analysis/parser/RawTextParser/src/test/resources/cc_projects/project_4.cc.json
