+++ analysis/parser/RawTextParser/src/test/resources/sampleproject/spaces/spaces_3.xyz
