+++ analysis/parser/RawTextParser/src/test/resources/sampleproject/spaces/spaces_4.xyz
