+++ analysis/parser/RawTextParser/src/test/resources/sampleproject/tabs.xyz
