+++ analysis/tools/ccsh/build.gradle
@@ -31,7 +31,7 @@ dependencies {
     def edgeFilter = project(':filter:EdgeFilter')
     def validationTool = project(':tools:ValidationTool')
     def csvExporter = project(':export:CSVExporter')
-    def indentationLevelParser = project(':import:IndentationLevelParser')
+    def rawTextParser = project(':parser:RawTextParser')
 
     // first compile is for dependency in main, testCompile is so our test suite can find all other tests
     compile codeMaatImporter; testCompile codeMaatImporter.sourceSets.test.output
@@ -48,7 +48,7 @@ dependencies {
     compile validationTool; testCompile validationTool.sourceSets.test.output
     compile csvExporter; testCompile csvExporter.sourceSets.test.output
     compile structureModifier; testCompile structureModifier.sourceSets.test.output
-    compile indentationLevelParser; testCompile indentationLevelParser.sourceSets.test.output
+    compile rawTextParser; testCompile rawTextParser.sourceSets.test.output
 
     compile group: 'info.picocli', name: 'picocli', version: picocli_version
